Add support for cucumber 3
