+++ CHANGELOG.md
@@ -1,20 +1,24 @@
+### 3.0.0 (Aug-12-2017)
+
+* Support for Cucumber 3
+* Use `cucumber-html-reporter@2.0.3` for < Cucumber@3
 
 ### 2.0.3 (Jul-17-2017)
 
-* Escape HTML on step name & fix <p> and <div> order for block 'scenario-container' [PR#97](https://github.com/gkushang/cucumber-html-reporter/pull/97). Resolve Issue [#52](https://github.com/gkushang/cucumber-html-reporter/issues/52) 
+* Escape HTML on step name & fix <p> and <div> order for block 'scenario-container' [PR#97](https://github.com/gkushang/cucumber-html-reporter/pull/97). Resolve Issue [#52](https://github.com/gkushang/cucumber-html-reporter/issues/52)
 
 
 ### 2.0.2 (Jul-10-2017)
 
 * Support scoped packages [PR#93](https://github.com/gkushang/cucumber-html-reporter/pull/93)
 
-* Fix issue[#85](https://github.com/gkushang/cucumber-html-reporter/issues/85) with Screenshot attachments with Selenium 3 & Cucumber 2 [PR#95](https://github.com/gkushang/cucumber-html-reporter/pull/95) 
+* Fix issue[#85](https://github.com/gkushang/cucumber-html-reporter/issues/85) with Screenshot attachments with Selenium 3 & Cucumber 2 [PR#95](https://github.com/gkushang/cucumber-html-reporter/pull/95)
 
 
 ### 2.0.1 (Jul-07-2017)
 
 * Save screenshots into custom directory [PR#88](https://github.com/gkushang/cucumber-html-reporter/pull/88)
-        
+
 * Added support for JSON attachments [PR#89](https://github.com/gkushang/cucumber-html-reporter/pull/89)
 
         Based on MIME type of the attachment. Payload is expected to be base64-encoded (this is based on the existing behaviour of the Cucumber Reports plugin for Jenkins).
@@ -26,13 +30,13 @@
 
 ##### Support for Cucumber 2
 
-* Cucumber 2 [PR#81](https://github.com/gkushang/cucumber-html-reporter/pull/81). Resolves Issues [#73](https://github.com/gkushang/cucumber-html-reporter/issues/73),[#72](https://github.com/gkushang/cucumber-html-reporter/issues/72), [#61](https://github.com/gkushang/cucumber-html-reporter/issues/61), [#55](https://github.com/gkushang/cucumber-html-reporter/issues/55) 
+* Cucumber 2 [PR#81](https://github.com/gkushang/cucumber-html-reporter/pull/81). Resolves Issues [#73](https://github.com/gkushang/cucumber-html-reporter/issues/73),[#72](https://github.com/gkushang/cucumber-html-reporter/issues/72), [#61](https://github.com/gkushang/cucumber-html-reporter/issues/61), [#55](https://github.com/gkushang/cucumber-html-reporter/issues/55)
     * Duration is reported as Milliseconds
     * Attachments are now plain text without encoding
     * DRY the templates and HTML scripts
     * Run Travis-CI with Node@7
     * Disable the Strict mode to test pending/undefined steps scenarios
-    
+
 * Use `cucumber-html-reporter@0.5.0` for < Cucumber@2
 
 ### 0.5.0 (Jun-01-2017)
@@ -96,14 +100,14 @@
 
 ##### Fix
 
-* Sanitize Screenshot filename [Issue#45](https://github.com/gkushang/cucumber-html-reporter/issues/45) [PR#46](https://github.com/gkushang/cucumber-html-reporter/pull/46) 
+* Sanitize Screenshot filename [Issue#45](https://github.com/gkushang/cucumber-html-reporter/issues/45) [PR#46](https://github.com/gkushang/cucumber-html-reporter/pull/46)
 
 
 ### 0.3.6 (Dec-06-2016)
 
 ##### Fix
 
-* Failure in Before hook should fail the Feature/Scenario, Add slice to rest of the scenario pie charts [PR#44](https://github.com/gkushang/cucumber-html-reporter/pull/44) 
+* Failure in Before hook should fail the Feature/Scenario, Add slice to rest of the scenario pie charts [PR#44](https://github.com/gkushang/cucumber-html-reporter/pull/44)
 
 
 ### 0.3.5 (Nov-29-2016)
@@ -111,7 +115,7 @@
 ##### Enhancements
 
 * Colors
- * Making labels & colors consistent on report [PR#42](https://github.com/gkushang/cucumber-html-reporter/pull/42) 
+ * Making labels & colors consistent on report [PR#42](https://github.com/gkushang/cucumber-html-reporter/pull/42)
 
 
 ### 0.3.4 (Nov-28-2016)
@@ -119,7 +123,7 @@
 ##### Enhancements
 
 * Step Duration
- * light gray the step duration to distinguish from the GWT Step description 
+ * light gray the step duration to distinguish from the GWT Step description
 
 
 ### 0.3.3 (Nov-28-2016)
@@ -199,14 +203,14 @@ Default: `undefined`
 ##### Enhancements
 
 * Option to add custom name to the project & adds footer: [PR#28](https://github.com/gkushang/cucumber-html-reporter/pull/28)
-  
-  
+
+
 ### 0.2.13 (Sep-27-2016)
 
 ##### Enhancements
 
 * Show time taken by each steps to complete the execution on report: [PR#17](https://github.com/gkushang/cucumber-html-reporter/pull/17) & [PR#27](https://github.com/gkushang/cucumber-html-reporter/pull/27)
-    
+
 
 ### 0.2.12 (Sep-27-2016)
 
@@ -244,14 +248,14 @@ Default: `undefined`
 
 ##### BugFix
 
-* Fixed the issue when report was breaking with the Cucumber's Doc String: [Issue#14](https://github.com/gkushang/cucumber-html-reporter/issues/14) 
+* Fixed the issue when report was breaking with the Cucumber's Doc String: [Issue#14](https://github.com/gkushang/cucumber-html-reporter/issues/14)
 
 
 ### 0.2.8 (Aug-30-2016)
 
 ##### Enhancements
 
-* Generate consolidated report from multiple JSON files: [PR#12](https://github.com/gkushang/cucumber-html-reporter/pull/12) 
+* Generate consolidated report from multiple JSON files: [PR#12](https://github.com/gkushang/cucumber-html-reporter/pull/12)
 
     * Provide the path of `jsonDir` to generate consolidated report,
 
@@ -303,25 +307,25 @@ var options = {
 
 ##### BugFix
 
-* Fixed the issue where Error messages were not printing on the report. 
+* Fixed the issue where Error messages were not printing on the report.
 
 ### 0.2.5 (Jul-28-2016)
 
 ##### Enhancements
 
-* trim the text to be printed on report 
+* trim the text to be printed on report
 
 ### 0.2.4 (Jul-28-2016)
 
 ##### BugFix
 
-* https://github.com/mavdi/grunt-cucumberjs/issues/86 
+* https://github.com/mavdi/grunt-cucumberjs/issues/86
 
 ### 0.2.3 (Jul-26-2016)
 
 ##### Enhancements
 
-* Published https://github.com/gkushang/cucumber-html-reporter/pull/10 Set charset as utf-8 
+* Published https://github.com/gkushang/cucumber-html-reporter/pull/10 Set charset as utf-8
 
 ### 0.2.2 (Jul-21-2016)
 
@@ -336,30 +340,30 @@ var options = {
 * Recursively create HTML report directory if does not exists
 * Remove outdated chai-fs depedency and use chai-should assertions
 * Lighter the background color or Scenario attachments
- 
+
 
 ### 0.2.0 (Jul-10-2016)
 
 ##### Support for Cucumber@1.2.0 version
 
 * Screenshot attachment support for Cucumber release >= @1.2.0 (https://github.com/cucumber/cucumber-js/blob/master/CHANGELOG.md#bug-fixes-1)
- 
+
 
 ### 0.1.6 (Jul-07-2016)
 
 ##### Enhancements
 
-* Format feature descriptions on report 
+* Format feature descriptions on report
 * Add overflow scroll bar for the bigger data-table
 * print error messaged in the pre
- 
+
 
 ### 0.1.5 (Jul-05-2016)
 
 ##### Enhancements
 
 * Show feature description on report
- 
+
 * Updated README
 
 ##### Bug fixes
@@ -372,7 +376,7 @@ var options = {
 ##### Enhancements
 
 * Using `path` instead of separators to make platform agnostic
- 
+
 * Updated tests `hooks`
 
 * README updated with the instructions on how to integrate reporter to the cucumber hooks
