+++ lib/reporter.js
@@ -212,13 +212,13 @@ var generateReport = function (options) {
                     if (step.embeddings !== undefined) {
                         step.embeddings.forEach(function (embedding) {
 
-                            if (embedding.mime_type === 'text/plain') {
+                            if (embedding.media.type === 'text/plain') {
                                 if (!step.text) {
                                     step.text = embedding.data;
                                 } else {
                                     step.text = step.text.concat('<br>' + embedding.data);
                                 }
-                            } else if (embedding.mime_type === 'application/json') {
+                            } else if (embedding.media.type === 'application/json') {
                                 var decoded = new Buffer(embedding.data, 'base64').toString('ascii');
 
                                 if (!step.text) {
@@ -226,7 +226,7 @@ var generateReport = function (options) {
                                 } else {
                                     step.text = step.text.concat('<br>' + decoded);
                                 }
-                            } else if (embedding.mime_type === 'image/png') {
+                            } else if (embedding.media.type === 'image/png') {
                                 step.image = 'data:image/png;base64,' + embedding.data;
 
                                 if ((options.storeScreenshots && options.storeScreenshots === true) ||
