+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cucumber-html-reporter",
-  "version": "2.0.3",
+  "version": "3.0.0",
   "description": "Generates Cucumber HTML reports in three different themes",
   "main": "index.js",
   "author": {
@@ -50,9 +50,10 @@
     }
   ],
   "scripts": {
-    "test": "npm run features && npm run clean",
+    "test": "npm run run:cucumber ; npm run create:reports && npm run clean",
     "clean": "rm -rf test/report/*.html test/report/*.json test/report/screenshot screenshots",
-    "features": "node node_modules/cucumber/bin/cucumber test/features/ -r test/features/step_definitions -f json:test/report/cucumber_report.json --no-strict"
+    "create:reports": "node test/createHtmlReports.js",
+    "run:cucumber": "node node_modules/cucumber/bin/cucumber test/features/ -r test/features/step_definitions -f json:test/report/cucumber_report.json --no-strict"
   },
   "repository": {
     "type": "git",
@@ -81,6 +82,6 @@
   },
   "devDependencies": {
     "chai": "^3.5.0",
-    "cucumber": "^2.3.0"
+    "cucumber": "^3.0.0"
   }
 }
